리뷰대응: NoticeType 위치 이동; domain -> common
