+++ src/main/java/com/LetMeDoWith/LetMeDoWith/infrastructure/notice/converter/NoticeTypeConverter.java
@@ -1,7 +1,7 @@
 package com.LetMeDoWith.LetMeDoWith.infrastructure.notice.converter;
 
 import com.LetMeDoWith.LetMeDoWith.common.converter.AbstractCombinedEnumConverter;
-import com.LetMeDoWith.LetMeDoWith.domain.notice.Notice.NoticeType;
+import com.LetMeDoWith.LetMeDoWith.common.enums.notice.NoticeType;
 import jakarta.persistence.Converter;
 
 @Converter(autoApply = true)
